commit6

+++ tasks/seq/vetoshnikova_k_convex_hull_bin_image_components/perf_tests/main.cpp
@@ -7,14 +7,14 @@
 #include "seq/vetoshnikova_k_convex_hull_bin_image_components/include/ops_seq.hpp"
 
 TEST(vetoshnikova_k_hull_bin_image_seq, test_pipeline_run) {
-  int w = 7;
-  int h = 10000;
+  uint32_t w = 7;
+  uint32_t h = 10000;
 
   // Create data
   std::vector<int> out(200);
-  std::vector<int8_t> b = {1, 1, 0, 0, 0, 0, 0};
-  std::vector<int8_t> a = {0, 0, 0, 0, 0, 0, 0};
-  std::vector<std::vector<int8_t>> in(h);
+  std::vector<uint8_t> b = {1, 1, 0, 0, 0, 0, 0};
+  std::vector<uint8_t> a = {0, 0, 0, 0, 0, 0, 0};
+  std::vector<std::vector<uint8_t>> in(h);
   in.emplace_back(b);
   in.emplace_back(b);
   for (int i = 2; i < h; i++) {
@@ -24,7 +24,7 @@ TEST(vetoshnikova_k_hull_bin_image_seq, test_pipeline_run) {
 
   // Create TaskData
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
-  taskDataSeq->inputs.emplace_back(reinterpret_cast<int*>(in.data()));
+  taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(in.data()));
   taskDataSeq->inputs_count.emplace_back(in.size());
   taskDataSeq->outputs.emplace_back(reinterpret_cast<int*>(out.data()));
 
@@ -54,13 +54,13 @@ TEST(vetoshnikova_k_hull_bin_image_seq, test_pipeline_run) {
 }
 
 TEST(vetoshnikova_k_hull_bin_image_seq, test_task_run) {
-  int w = 7;
-  int h = 10000;
+  uint32_t w = 7;
+  uint32_t h = 10000;
 
   // Create data
   std::vector<int> out(200);
-  std::vector<int8_t> b = {1, 1, 0, 0, 0, 0, 0};
-  std::vector<std::vector<int8_t>> in(h);
+  std::vector<uint8_t> b = {1, 1, 0, 0, 0, 0, 0};
+  std::vector<std::vector<uint8_t>> in(h);
   for (int i = 0; i < h; i++) {
     in.emplace_back(b);
   }
@@ -68,7 +68,7 @@ TEST(vetoshnikova_k_hull_bin_image_seq, test_task_run) {
 
   // Create TaskData
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
-  taskDataSeq->inputs.emplace_back(reinterpret_cast<int*>(in.data()));
+  taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(in.data()));
   taskDataSeq->inputs_count.emplace_back(in.size());
   taskDataSeq->outputs.emplace_back(reinterpret_cast<int*>(out.data()));
